refactor: QuestionService QuestionResourceService 생성 - reuqest body 생성 로직 분리 AiService - 통신 로직 분리 S3Service - 파일 읽기 로직 분리 Entity - Builder패턴 클래스 단위로 이동
