+++ src/main/java/com/driveu/server/domain/question/application/QuestionService.java
@@ -1,9 +1,6 @@
 package com.driveu.server.domain.question.application;
 
 import com.amazonaws.services.kms.model.NotFoundException;
-import com.amazonaws.services.s3.AmazonS3Client;
-import com.amazonaws.services.s3.model.S3Object;
-import com.amazonaws.services.s3.model.S3ObjectInputStream;
 import com.driveu.server.domain.directory.dao.DirectoryRepository;
 import com.driveu.server.domain.directory.domain.Directory;
 import com.driveu.server.domain.question.dao.QuestionRepository;
@@ -12,29 +9,18 @@ import com.driveu.server.domain.question.domain.Question;
 import com.driveu.server.domain.question.dto.request.QuestionCreateRequest;
 import com.driveu.server.domain.question.dto.response.QuestionListResponse;
 import com.driveu.server.domain.question.dto.response.QuestionResponse;
-import com.driveu.server.domain.resource.dao.FileRepository;
-import com.driveu.server.domain.resource.dao.NoteRepository;
 import com.driveu.server.domain.resource.dao.ResourceDirectoryRepository;
 import com.driveu.server.domain.resource.dao.ResourceRepository;
-import com.driveu.server.domain.resource.domain.File;
-import com.driveu.server.domain.resource.domain.Note;
 import com.driveu.server.domain.resource.domain.Resource;
 import com.driveu.server.domain.resource.domain.ResourceDirectory;
 import com.driveu.server.infra.ai.AiService;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.jetbrains.annotations.NotNull;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.io.ByteArrayResource;
 import org.springframework.http.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Paths;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -49,13 +35,8 @@ public class QuestionService {
     private final QuestionResourceRepository questionResourceRepository;
     private final ResourceRepository resourceRepository;
     private final ResourceDirectoryRepository resourceDirectoryRepository;
-    private final NoteRepository noteRepository;
     private final AiService aiService;
-    private final AmazonS3Client amazonS3Client;
-    private final FileRepository fileRepository;
-
-    @Value("${spring.cloud.aws.s3.bucket}")
-    private String bucketName;
+    private final QuestionResourceService questionResourceService;
 
     @Transactional
     public QuestionResponse createQuestion(Long directoryId, List<QuestionCreateRequest> requestList) {
@@ -81,7 +62,7 @@ public class QuestionService {
         int version = getVersion(existingSameQuestions);
 
         // resource type에 따라 파일을 추출해서 multipart/form-data 데이터 형식으로 만듦
-        MultiValueMap<String, Object> requestBody = createRequestBody(requestList);
+        MultiValueMap<String, Object> requestBody = questionResourceService.createRequestBody(requestList);
         String aiResponse = aiService.generateQuestion(requestBody);
 
         Question question = Question.of(title, version, aiResponse);
@@ -98,67 +79,6 @@ public class QuestionService {
         questionRepository.flush();
 
         return QuestionResponse.fromEntity(savedQuestion);
-    }
-
-    private @NotNull MultiValueMap<String, Object> createRequestBody(List<QuestionCreateRequest> requestList) {
-        MultiValueMap<String, Object> body = new LinkedMultiValueMap<>();
-
-        for (QuestionCreateRequest request : requestList) {
-            if (request.getType().equals("FILE")){
-                // 파일에 저장된 s3Path로 S3에서 가져오기
-                addFileFromS3(request, body);
-
-            } else if(request.getType().equals("NOTE")) {
-                // 노트 컨텐츠로 파일 만들기
-                addNote(request, body);
-            }
-            else {
-                throw new IllegalArgumentException("잘못된 resource type 입니다.");
-            }
-        }
-        return body;
-    }
-
-    private void addNote(QuestionCreateRequest request, MultiValueMap<String, Object> body) {
-        Note note = noteRepository.findById(request.getResourceId())
-                .orElseThrow(() -> new EntityNotFoundException("Note not found: " + request.getResourceId()));
-
-        String markdown = note.getContent();
-
-        // String → ByteArrayResource (가짜 파일)
-        ByteArrayResource fileResource = new ByteArrayResource(
-                markdown.getBytes(StandardCharsets.UTF_8)
-        ) {
-            @Override
-            public String getFilename() {
-                return "note-" + request.getResourceId() + ".md";
-            }
-        };
-
-        // 같은 폼 필드명("files")에 여러 개를 add하면, 서버 쪽에서는 배열로 받는다.
-        body.add("files", fileResource);
-    }
-
-    private void addFileFromS3(QuestionCreateRequest request, MultiValueMap<String, Object> body) {
-        File file = fileRepository.findById(request.getResourceId())
-                .orElseThrow(() -> new EntityNotFoundException("File not found: " + request.getResourceId()));
-
-        String s3Path = file.getS3Path();
-        String filename = Paths.get(s3Path).getFileName().toString();
-
-        S3Object s3Object = amazonS3Client.getObject(bucketName, s3Path);
-        try (S3ObjectInputStream is = s3Object.getObjectContent()) {
-            byte[] bytes = is.readAllBytes();
-            ByteArrayResource fileResource = new ByteArrayResource(bytes) {
-                @Override
-                public String getFilename() {
-                    return filename;
-                }
-            };
-            body.add("files", fileResource);
-        } catch (IOException e) {
-            throw new RuntimeException("S3에서 파일 읽기 실패: " + s3Path, e);
-        }
     }
 
     private @NotNull String createTitle(List<QuestionCreateRequest> requestList, Directory directory) {
@@ -188,7 +108,7 @@ public class QuestionService {
         return version;
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public QuestionResponse getQuestionById(Long questionId) {
         Question question = questionRepository.findById(questionId)
                 .orElseThrow(() -> new NotFoundException("Question not found"));
@@ -196,7 +116,7 @@ public class QuestionService {
         return QuestionResponse.fromEntity(question);
     }
 
-    @Transactional
+    @Transactional(readOnly = true)
     public List<QuestionListResponse> getQuestionsByUserSemester(Long userSemesterId) {
 
         // 1) 해당 학기에 속한 모든 디렉토리 조회
